pyproject.toml
[project]
name = "langflow-runtime"
version = "0.1.0"
description = "Agentic AI Workshop with Langflow"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langflow >= 1.2.0",
    "streamlit >= 1.32.0",
    "python-dotenv >= 1.0.0",
    "requests >= 2.31.0"
]